# cli/internal/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn17AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn2AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn3AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn16AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn8AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn18AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn9AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn11AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn12AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn14AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn6AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn7AllowedHeaders = map[string]string{
		"POST": "Authorization",
	}
	rn13AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn19AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/"

			if l := len("/api/"); len(elem) >= l && elem[0:l] == "/api/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "agent/"

				if l := len("agent/"); len(elem) >= l && elem[0:l] == "agent/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'o': // Prefix: "onboarding"

					if l := len("onboarding"); len(elem) >= l && elem[0:l] == "onboarding" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "POST":
							s.handleStartOnboardingRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn17AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "run_id"
						// Match until "/"
						idx := strings.IndexByte(elem, '/')
						if idx < 0 {
							idx = len(elem)
						}
						args[0] = elem[:idx]
						elem = elem[idx:]

						if len(elem) == 0 {
							switch r.Method {
							case "GET":
								s.handleGetOnboardingRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: rn2AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}
						switch elem[0] {
						case '/': // Prefix: "/complete"

							if l := len("/complete"); len(elem) >= l && elem[0:l] == "/complete" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "POST":
									s.handleCompleteOnboardingRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "POST",
										allowedHeaders: rn3AllowedHeaders,
										acceptPost:     "application/json",
										acceptPatch:    "",
									})
								}

								return
							}

						}

					}

				case 's': // Prefix: "search"

					if l := len("search"); len(elem) >= l && elem[0:l] == "search" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleSearchAgentRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn16AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'c': // Prefix: "content/"

				if l := len("content/"); len(elem) >= l && elem[0:l] == "content/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListContentRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn8AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case 's': // Prefix: "submit"
					origElem := elem
					if l := len("submit"); len(elem) >= l && elem[0:l] == "submit" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleSubmitContentRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn18AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				}
				// Param: "content_id"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetContentRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn9AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'j': // Prefix: "jobs/"

				if l := len("jobs/"); len(elem) >= l && elem[0:l] == "jobs/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "job_id"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetJobRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn11AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'n': // Prefix: "news/items"

				if l := len("news/items"); len(elem) >= l && elem[0:l] == "news/items" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListNewsItemsRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn12AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'm': // Prefix: "mark-read"
						origElem := elem
						if l := len("mark-read"); len(elem) >= l && elem[0:l] == "mark-read" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleMarkNewsItemsReadRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn14AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

						elem = origElem
					}
					// Param: "news_item_id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleGetNewsItemRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn6AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/convert-to-article"

						if l := len("/convert-to-article"); len(elem) >= l && elem[0:l] == "/convert-to-article" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleConvertNewsItemToArticleRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn7AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				}

			case 's': // Prefix: "scrapers/"

				if l := len("scrapers/"); len(elem) >= l && elem[0:l] == "scrapers/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListSourcesRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn13AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case 's': // Prefix: "subscribe"

					if l := len("subscribe"); len(elem) >= l && elem[0:l] == "subscribe" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleSubscribeSourceRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn19AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/"

			if l := len("/api/"); len(elem) >= l && elem[0:l] == "/api/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "agent/"

				if l := len("agent/"); len(elem) >= l && elem[0:l] == "agent/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'o': // Prefix: "onboarding"

					if l := len("onboarding"); len(elem) >= l && elem[0:l] == "onboarding" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "POST":
							r.name = StartOnboardingOperation
							r.summary = "Start Onboarding"
							r.operationID = "startOnboarding"
							r.operationGroup = ""
							r.pathPattern = "/api/agent/onboarding"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "run_id"
						// Match until "/"
						idx := strings.IndexByte(elem, '/')
						if idx < 0 {
							idx = len(elem)
						}
						args[0] = elem[:idx]
						elem = elem[idx:]

						if len(elem) == 0 {
							switch method {
							case "GET":
								r.name = GetOnboardingOperation
								r.summary = "Get Onboarding"
								r.operationID = "getOnboarding"
								r.operationGroup = ""
								r.pathPattern = "/api/agent/onboarding/{run_id}"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}
						switch elem[0] {
						case '/': // Prefix: "/complete"

							if l := len("/complete"); len(elem) >= l && elem[0:l] == "/complete" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "POST":
									r.name = CompleteOnboardingOperation
									r.summary = "Complete Onboarding"
									r.operationID = "completeOnboarding"
									r.operationGroup = ""
									r.pathPattern = "/api/agent/onboarding/{run_id}/complete"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						}

					}

				case 's': // Prefix: "search"

					if l := len("search"); len(elem) >= l && elem[0:l] == "search" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = SearchAgentOperation
							r.summary = "Search Agent"
							r.operationID = "searchAgent"
							r.operationGroup = ""
							r.pathPattern = "/api/agent/search"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'c': // Prefix: "content/"

				if l := len("content/"); len(elem) >= l && elem[0:l] == "content/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListContentOperation
						r.summary = "List content items"
						r.operationID = "listContent"
						r.operationGroup = ""
						r.pathPattern = "/api/content/"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case 's': // Prefix: "submit"
					origElem := elem
					if l := len("submit"); len(elem) >= l && elem[0:l] == "submit" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = SubmitContentOperation
							r.summary = "Submit a one-off URL for processing"
							r.operationID = "submitContent"
							r.operationGroup = ""
							r.pathPattern = "/api/content/submit"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				}
				// Param: "content_id"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetContentOperation
						r.summary = "Get content details"
						r.operationID = "getContent"
						r.operationGroup = ""
						r.pathPattern = "/api/content/{content_id}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			case 'j': // Prefix: "jobs/"

				if l := len("jobs/"); len(elem) >= l && elem[0:l] == "jobs/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "job_id"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetJobOperation
						r.summary = "Get Job"
						r.operationID = "getJob"
						r.operationGroup = ""
						r.pathPattern = "/api/jobs/{job_id}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			case 'n': // Prefix: "news/items"

				if l := len("news/items"); len(elem) >= l && elem[0:l] == "news/items" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListNewsItemsOperation
						r.summary = "List visible news items"
						r.operationID = "listNewsItems"
						r.operationGroup = ""
						r.pathPattern = "/api/news/items"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'm': // Prefix: "mark-read"
						origElem := elem
						if l := len("mark-read"); len(elem) >= l && elem[0:l] == "mark-read" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = MarkNewsItemsReadOperation
								r.summary = "Mark visible news items as read"
								r.operationID = "markNewsItemsRead"
								r.operationGroup = ""
								r.pathPattern = "/api/news/items/mark-read"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

						elem = origElem
					}
					// Param: "news_item_id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = GetNewsItemOperation
							r.summary = "Get one news item"
							r.operationID = "getNewsItem"
							r.operationGroup = ""
							r.pathPattern = "/api/news/items/{news_item_id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/convert-to-article"

						if l := len("/convert-to-article"); len(elem) >= l && elem[0:l] == "/convert-to-article" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = ConvertNewsItemToArticleOperation
								r.summary = "Convert one news item into article content"
								r.operationID = "convertNewsItemToArticle"
								r.operationGroup = ""
								r.pathPattern = "/api/news/items/{news_item_id}/convert-to-article"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			case 's': // Prefix: "scrapers/"

				if l := len("scrapers/"); len(elem) >= l && elem[0:l] == "scrapers/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListSourcesOperation
						r.summary = "List Scraper Configs"
						r.operationID = "listSources"
						r.operationGroup = ""
						r.pathPattern = "/api/scrapers/"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case 's': // Prefix: "subscribe"

					if l := len("subscribe"); len(elem) >= l && elem[0:l] == "subscribe" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = SubscribeSourceOperation
							r.summary = "Subscribe To Feed"
							r.operationID = "subscribeSource"
							r.operationGroup = ""
							r.pathPattern = "/api/scrapers/subscribe"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			}

		}
	}
	return r, false
}