# cli/openapi/agent-openapi.json
{
  "components": {
    "schemas": {
...
      "AgentOnboardingStartRequest": {
        "description": "Simplified async onboarding start request.",
        "properties": {
          "brief": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Brief",
            "type": "string"
          },
          "preferences": {
            "additionalProperties": true,
            "nullable": true,
            "title": "Preferences",
            "type": "object"
          },
          "seed_feeds": {
            "items": {
              "type": "string"
            },
            "title": "Seed Feeds",
            "type": "array"
          },
          "seed_urls": {
            "items": {
              "type": "string"
            },
            "title": "Seed Urls",
            "type": "array"
          }
        },
        "required": [
          "brief"
        ],
        "title": "AgentOnboardingStartRequest",
        "type": "object"
      },
      "AgentOnboardingStartResponse": {
        "description": "Simplified async onboarding start response.",
        "properties": {
          "job_id": {
            "nullable": true,
            "title": "Job Id",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "AgentOnboardingStartResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Filtered machine-oriented API contract for the standalone newsly-agent CLI.",
    "title": "Newsly Agent CLI API",
    "version": "2.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/agent/onboarding": {
      "post": {
        "description": "Start simplified async onboarding.",
        "operationId": "startOnboarding",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentOnboardingStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOnboardingStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Onboarding",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/agent/onboarding/{run_id}": {
      "get": {
        "description": "Return onboarding run status.",
        "operationId": "getOnboarding",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingDiscoveryStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Onboarding",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/agent/onboarding/{run_id}/complete": {
      "post": {
        "description": "Complete onboarding from simplified selections.",
        "operationId": "completeOnboarding",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentOnboardingCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Complete Onboarding Api Agent Onboarding  Run Id  Complete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Onboarding",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/agent/search": {
      "post": {
        "description": "Search external/provider-backed sources for the agent CLI.",
        "operationId": "searchAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Agent",
        "tags": [
          "search"
        ]
      }
    },
    "/api/content/": {
      "get": {
        "description": "Retrieve a list of content items with optional filtering by content type and date. Supports multiple content types via repeated query parameters (e.g., ?content_type=article&content_type=podcast). Supports cursor-based pagination for efficient loading.",
        "operationId": "listContent",
        "parameters": [
          {
            "description": "Filter by content type(s). Can be specified multiple times for multiple types (article/podcast/news)",
            "in": "query",
            "name": "content_type",
            "required": false,
            "schema": {
              "description": "Filter by content type(s). Can be specified multiple times for multiple types (article/podcast/news)",
              "items": {
                "type": "string"
              },
              "nullable": true,
              "title": "Content Type",
              "type": "array"
            }
          },
          {
            "description": "Filter by date (YYYY-MM-DD format)",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "description": "Filter by date (YYYY-MM-DD format)",
              "nullable": true,
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "description": "Filter by read status (all/read/unread)",
            "in": "query",
            "name": "read_filter",
            "required": false,
            "schema": {
              "default": "all",
              "description": "Filter by read status (all/read/unread)",
              "pattern": "^(all|read|unread)$",
              "title": "Read Filter",
              "type": "string"
            }
          },
          {
            "description": "Pagination cursor for next page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "description": "Pagination cursor for next page",
              "nullable": true,
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "description": "Number of items per page (max 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Number of items per page (max 100)",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Whether to include the available_dates metadata used by filterable list UIs. Disable for feed surfaces that do not show date filters.",
            "in": "query",
            "name": "include_available_dates",
            "required": false,
            "schema": {
              "default": true,
              "description": "Whether to include the available_dates metadata used by filterable list UIs. Disable for feed surfaces that do not show date filters.",
              "title": "Include Available Dates",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List content items",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/submit": {
      "post": {
        "description": "Submit article or podcast URLs for processing. Only http/https URLs are accepted.",
        "operationId": "submitContent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentSubmissionResponse"
                }
              }
            },
            "description": "Existing content matched and reused"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentSubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit a one-off URL for processing",
        "tags": [
          "content"
        ]
      }
    },
...
  },
...
}